[mod/#230]getMyRecentReviewStats() 도메인에 따라 이동
