+++ app/src/main/java/com/smashing/app/data/remote/datasource/api/MyRemoteDataSource.kt
@@ -2,6 +2,7 @@ package com.smashing.app.data.remote.datasource.api
 import com.smashing.app.data.remote.dto.BaseResponse
 import com.smashing.app.data.remote.dto.my.AddSportProfileRequest
 import com.smashing.app.data.remote.dto.my.GetMyProfileResponse
+import com.smashing.app.data.remote.dto.my.GetMyRecentReviewStatsResponse
 import com.smashing.app.data.remote.dto.my.GetMyTierProfileResponse
 import com.smashing.app.data.remote.dto.my.MyProfileSwitchRequest
 
@@ -11,4 +12,6 @@ interface MyRemoteDataSource {
 
     suspend fun addSportProfile(request: AddSportProfileRequest): BaseResponse<Unit?>
     suspend fun putActiveMyProfile(request: MyProfileSwitchRequest): BaseResponse<Unit?>
+    suspend fun getMyRecentReviewStats(): BaseResponse<GetMyRecentReviewStatsResponse>
+
 }
