+++ app/src/main/java/com/smashing/app/data/remote/datasource/api/ReviewRemoteDataSource.kt
@@ -4,7 +4,6 @@ import com.smashing.app.data.remote.dto.BaseResponse
 import com.smashing.app.data.remote.dto.cursor.CursorDto
 import com.smashing.app.data.remote.dto.review.GetReviewDetailResponse
 import com.smashing.app.data.remote.dto.review.GetMyRecentReviewListResponse
-import com.smashing.app.data.remote.dto.my.GetMyRecentReviewStatsResponse
 import com.smashing.app.data.remote.dto.review.GetUserRecentReviewListResponse
 
 interface ReviewRemoteDataSource {
@@ -24,6 +23,4 @@ interface ReviewRemoteDataSource {
         cursor: String?,
         size: Int?
     ): BaseResponse<CursorDto<GetMyRecentReviewListResponse>>
-
-    suspend fun getMyRecentReviewStats(): BaseResponse<GetMyRecentReviewStatsResponse>
 }
