+++ app/src/main/java/com/smashing/app/data/remote/datasource/impl/ReviewRemoteDataSourceImpl.kt
@@ -5,7 +5,6 @@ import com.smashing.app.data.remote.dto.BaseResponse
 import com.smashing.app.data.remote.dto.cursor.CursorDto
 import com.smashing.app.data.remote.dto.review.GetReviewDetailResponse
 import com.smashing.app.data.remote.dto.review.GetMyRecentReviewListResponse
-import com.smashing.app.data.remote.dto.my.GetMyRecentReviewStatsResponse
 import com.smashing.app.data.remote.dto.review.GetUserRecentReviewListResponse
 import com.smashing.app.data.remote.service.ReviewService
 import javax.inject.Inject
@@ -41,7 +40,4 @@ class ReviewRemoteDataSourceImpl @Inject constructor(
             size = size,
         )
 
-    override suspend fun getMyRecentReviewStats(): BaseResponse<GetMyRecentReviewStatsResponse> {
-        return reviewService.getMyRecentReviewStats()
-    }
 }
