+++ app/src/main/java/com/smashing/app/data/remote/service/MyService.kt
@@ -3,6 +3,7 @@ package com.smashing.app.data.remote.service
 import com.smashing.app.data.remote.dto.BaseResponse
 import com.smashing.app.data.remote.dto.my.AddSportProfileRequest
 import com.smashing.app.data.remote.dto.my.GetMyProfileResponse
+import com.smashing.app.data.remote.dto.my.GetMyRecentReviewStatsResponse
 import com.smashing.app.data.remote.dto.my.GetMyTierProfileResponse
 import com.smashing.app.data.remote.dto.my.MyProfileSwitchRequest
 import retrofit2.http.Body
@@ -26,4 +27,9 @@ interface MyService {
 
     @GET("/api/v1/users/me/profiles/tier")
     suspend fun getMyTierProfile(): BaseResponse<GetMyTierProfileResponse>
+
+    @GET("/api/v1/users/me/reviews/summary")
+    suspend fun getMyRecentReviewStats(): BaseResponse<GetMyRecentReviewStatsResponse>
+
+
 }
