+++ app/src/main/java/com/smashing/app/data/repository/impl/ReviewRepositoryImpl.kt
@@ -1,12 +1,10 @@
 package com.smashing.app.data.repository.impl
 
 import com.smashing.app.core.util.suspendRunCatching
-import com.smashing.app.data.mapper.my.toGameReviewResult
 import com.smashing.app.data.mapper.review.toGameReviewList
 import com.smashing.app.data.mapper.review.toReviewDetail
 import com.smashing.app.data.model.cursor.CursorPage
 import com.smashing.app.data.model.review.GameReview
-import com.smashing.app.data.model.review.GameReviewResult
 import com.smashing.app.data.model.review.ReviewDetail
 import com.smashing.app.data.remote.datasource.api.ReviewRemoteDataSource
 import com.smashing.app.data.remote.dto.requireData
@@ -51,10 +49,4 @@ class ReviewRepositoryImpl @Inject constructor(
         ).requireData().toGameReviewList()
     }
 
-    override suspend fun getUserRecentReviewStats(
-    ): Result<GameReviewResult> = suspendRunCatching {
-        reviewRemoteDataSource.getMyRecentReviewStats()
-            .requireData()
-            .toGameReviewResult()
-    }
 }
