+++ app/src/main/java/com/smashing/app/data/repository/impl/UserRepositoryImpl.kt
@@ -4,8 +4,8 @@ import com.smashing.app.core.util.suspendRunCatching
 import com.smashing.app.data.local.datasource.api.LocalUserDataSource
 import com.smashing.app.data.mapper.user.toGameReviewResult
 import com.smashing.app.data.mapper.user.toUserProfileInfo
-import com.smashing.app.data.model.review.GameReviewResult
 import com.smashing.app.data.model.profile.user.UserProfileInfo
+import com.smashing.app.data.model.review.GameReviewResult
 import com.smashing.app.data.remote.datasource.api.UserRemoteDataSource
 import com.smashing.app.data.remote.dto.requireData
 import com.smashing.app.data.repository.api.UserRepository
@@ -42,4 +42,5 @@ class UserRepositoryImpl @Inject constructor(
         suspendRunCatching{
             userRemoteDataSource.getUserRecentReviewStats(userId, sportCode).requireData().toGameReviewResult()
         }
+
 }
